requirements.txt
streamlit
pandas
numpy
msoffcrypto-tool
matplotlib
plotly
seaborn
streamlit-aggrid
openpyxl
openpyxl.styles
dateutil.relativedelta
pygwalker
chart_studio
xlrd
openpyxl
openai
pillow
streamlit_pdf_viewer
lxml